Add panic functionality to our codebase
Properly defines a manic method so that we can use it throughout
the code to allow us to properly display unrecoverable errors on
our LEDs.
